+++ config/doguConfig.go
@@ -1,13 +1,15 @@
 package config
 
+import "github.com/cloudogu/ces-commons-lib/dogu"
+
 // DoguConfig represents a Dogu-specific configuration.
 type DoguConfig struct {
-	DoguName SimpleDoguName
+	DoguName dogu.SimpleName
 	Config
 }
 
 // CreateDoguConfig creates a new Dogu-specific configuration with the provided Dogu name and entries.
-func CreateDoguConfig(dogu SimpleDoguName, e Entries) DoguConfig {
+func CreateDoguConfig(dogu dogu.SimpleName, e Entries) DoguConfig {
 	return DoguConfig{
 		DoguName: dogu,
 		Config:   CreateConfig(e),