+++ dogu/doguVersionRegistry_test.go
@@ -3,8 +3,9 @@ package dogu
 import (
 	"context"
 	"fmt"
+	"github.com/cloudogu/ces-commons-lib/dogu"
+	cloudoguerrors "github.com/cloudogu/ces-commons-lib/errors"
 	"github.com/cloudogu/cesapp-lib/core"
-	cloudoguerrors "github.com/cloudogu/k8s-registry-lib/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
@@ -39,15 +40,15 @@ func TestNewDoguVersionRegistry(t *testing.T) {
 }
 
 func Test_versionRegistry_GetCurrent(t *testing.T) {
-	expectedDoguVersion := DoguVersion{
-		Name:    "cas",
+	expectedDoguVersion := dogu.SimpleNameVersion{
+		Name:    dogu.SimpleName("cas"),
 		Version: parseVersionStr(t, casVersionStr),
 	}
 	casRegistryCm := &corev1.ConfigMap{Data: map[string]string{"current": casVersionStr}}
 
 	type args struct {
 		ctx  context.Context
-		name SimpleDoguName
+		name dogu.SimpleName
 	}
 	casArgs := args{
 		ctx:  testCtx,
@@ -57,7 +58,7 @@ func Test_versionRegistry_GetCurrent(t *testing.T) {
 		name              string
 		configMapClientFn func(t *testing.T) configMapClient
 		args              args
-		want              DoguVersion
+		want              dogu.SimpleNameVersion
 		wantErr           assert.ErrorAssertionFunc
 	}{
 		{
@@ -81,7 +82,7 @@ func Test_versionRegistry_GetCurrent(t *testing.T) {
 				return configMapClientMock
 			},
 			args: casArgs,
-			want: DoguVersion{},
+			want: dogu.SimpleNameVersion{},
 			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
 				return assert.True(t, cloudoguerrors.IsGenericError(err), i) &&
 					assert.ErrorContains(t, err, "failed to get dogu descriptor config map for dogu \"cas\"", i)
@@ -96,7 +97,7 @@ func Test_versionRegistry_GetCurrent(t *testing.T) {
 				return configMapClientMock
 			},
 			args: casArgs,
-			want: DoguVersion{},
+			want: dogu.SimpleNameVersion{},
 			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
 				return assert.True(t, cloudoguerrors.IsNotFoundError(err), i) &&
 					assert.ErrorContains(t, err, "failed to get value for key \"current\" for dogu registry \"cas\"", i)
@@ -111,7 +112,7 @@ func Test_versionRegistry_GetCurrent(t *testing.T) {
 				return configMapClientMock
 			},
 			args: casArgs,
-			want: DoguVersion{},
+			want: dogu.SimpleNameVersion{},
 			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
 				return assert.True(t, cloudoguerrors.IsGenericError(err), i) &&
 					assert.ErrorContains(t, err, "failed to parse version \"abc\" for dogu \"cas\"", i)
@@ -134,12 +135,12 @@ func Test_versionRegistry_GetCurrent(t *testing.T) {
 }
 
 func Test_versionRegistry_GetCurrentOfAll(t *testing.T) {
-	expectedCasDoguVersion := DoguVersion{
-		Name:    "cas",
+	expectedCasDoguVersion := dogu.SimpleNameVersion{
+		Name:    dogu.SimpleName("cas"),
 		Version: parseVersionStr(t, casVersionStr),
 	}
-	expectedLdapDoguVersion := DoguVersion{
-		Name:    "ldap",
+	expectedLdapDoguVersion := dogu.SimpleNameVersion{
+		Name:    dogu.SimpleName("ldap"),
 		Version: parseVersionStr(t, ldapVersionStr),
 	}
 	casRegistryCm := &corev1.ConfigMap{Data: map[string]string{"current": casVersionStr}, ObjectMeta: metav1.ObjectMeta{Labels: casVersionRegistryLabelMap}}
@@ -153,7 +154,7 @@ func Test_versionRegistry_GetCurrentOfAll(t *testing.T) {
 		name              string
 		configMapClientFn func(t *testing.T) configMapClient
 		args              args
-		want              []DoguVersion
+		want              []dogu.SimpleNameVersion
 		wantErr           assert.ErrorAssertionFunc
 	}{
 		{
@@ -165,7 +166,7 @@ func Test_versionRegistry_GetCurrentOfAll(t *testing.T) {
 				return configMapClientMock
 			},
 			args:    args{ctx: testCtx},
-			want:    []DoguVersion{expectedCasDoguVersion, expectedLdapDoguVersion},
+			want:    []dogu.SimpleNameVersion{expectedCasDoguVersion, expectedLdapDoguVersion},
 			wantErr: assert.NoError,
 		},
 		{
@@ -179,7 +180,7 @@ func Test_versionRegistry_GetCurrentOfAll(t *testing.T) {
 				return configMapClientMock
 			},
 			args:    args{ctx: testCtx},
-			want:    []DoguVersion{expectedLdapDoguVersion},
+			want:    []dogu.SimpleNameVersion{expectedLdapDoguVersion},
 			wantErr: assert.NoError,
 		},
 		{
@@ -209,7 +210,7 @@ func Test_versionRegistry_GetCurrentOfAll(t *testing.T) {
 				return configMapClientMock
 			},
 			args: args{ctx: testCtx},
-			want: []DoguVersion{},
+			want: []dogu.SimpleNameVersion{},
 			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
 				return assert.True(t, cloudoguerrors.IsGenericError(err), i) &&
 					assert.ErrorContains(t, err, "failed to get some dogu versions: failed to parse version \"abc\" for dogu \"cas\": failed to parse major version abc: strconv.Atoi: parsing \"abc\": invalid syntax\nfailed to parse version \"abcd\" for dogu \"ldap\": failed to parse major version abcd: strconv.Atoi: parsing \"abcd\": invalid syntax")
@@ -227,9 +228,9 @@ func Test_versionRegistry_GetCurrentOfAll(t *testing.T) {
 				return configMapClientMock
 			},
 			args: args{ctx: testCtx},
-			want: []DoguVersion{
+			want: []dogu.SimpleNameVersion{
 				{
-					Name:    "ldap",
+					Name:    dogu.SimpleName("ldap"),
 					Version: parseVersionStr(t, "1.0.0"),
 				},
 			},
@@ -258,7 +259,7 @@ func Test_versionRegistry_IsEnabled(t *testing.T) {
 	casRegistryCmWithOutCurrent := &corev1.ConfigMap{Data: map[string]string{casVersionStr: readCasDoguStr(t)}, ObjectMeta: metav1.ObjectMeta{Labels: casVersionRegistryLabelMap}}
 	type args struct {
 		ctx         context.Context
-		doguVersion DoguVersion
+		doguVersion dogu.SimpleNameVersion
 	}
 	tests := []struct {
 		name              string
@@ -275,7 +276,9 @@ func Test_versionRegistry_IsEnabled(t *testing.T) {
 
 				return configMapClientMock
 			},
-			args:    args{ctx: testCtx, doguVersion: DoguVersion{"cas", parseVersionStr(t, casVersionStr)}},
+			args: args{ctx: testCtx, doguVersion: dogu.SimpleNameVersion{
+				Name:    dogu.SimpleName("cas"),
+				Version: parseVersionStr(t, casVersionStr)}},
 			want:    true,
 			wantErr: assert.NoError,
 		},
@@ -287,7 +290,9 @@ func Test_versionRegistry_IsEnabled(t *testing.T) {
 
 				return configMapClientMock
 			},
-			args:    args{ctx: testCtx, doguVersion: DoguVersion{"cas", parseVersionStr(t, casVersionStr)}},
+			args: args{ctx: testCtx, doguVersion: dogu.SimpleNameVersion{
+				Name:    dogu.SimpleName("cas"),
+				Version: parseVersionStr(t, casVersionStr)}},
 			want:    false,
 			wantErr: assert.NoError,
 		},
@@ -299,7 +304,9 @@ func Test_versionRegistry_IsEnabled(t *testing.T) {
 
 				return configMapClientMock
 			},
-			args:    args{ctx: testCtx, doguVersion: DoguVersion{"cas", parseVersionStr(t, "7.0.5.1-2")}},
+			args: args{ctx: testCtx, doguVersion: dogu.SimpleNameVersion{
+				Name:    dogu.SimpleName("cas"),
+				Version: parseVersionStr(t, "7.0.5.1-2")}},
 			want:    false,
 			wantErr: assert.NoError,
 		},
@@ -311,7 +318,9 @@ func Test_versionRegistry_IsEnabled(t *testing.T) {
 
 				return configMapClientMock
 			},
-			args: args{ctx: testCtx, doguVersion: DoguVersion{"cas", parseVersionStr(t, casVersionStr)}},
+			args: args{ctx: testCtx, doguVersion: dogu.SimpleNameVersion{
+				Name:    dogu.SimpleName("cas"),
+				Version: parseVersionStr(t, casVersionStr)}},
 			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
 				return assert.True(t, cloudoguerrors.IsGenericError(err), i) &&
 					assert.ErrorContains(t, err, "failed to get dogu descriptor config map for dogu \"cas\"")
@@ -338,11 +347,12 @@ func Test_versionRegistry_Enable(t *testing.T) {
 
 	type args struct {
 		ctx         context.Context
-		doguVersion DoguVersion
+		doguVersion dogu.SimpleNameVersion
 	}
 	casArgs := args{
-		ctx:         testCtx,
-		doguVersion: DoguVersion{Name: "cas", Version: parseVersionStr(t, casVersionStr)},
+		ctx: testCtx,
+		doguVersion: dogu.SimpleNameVersion{
+			Name: dogu.SimpleName("cas"), Version: parseVersionStr(t, casVersionStr)},
 	}
 	tests := []struct {
 		name              string
@@ -422,7 +432,7 @@ func Test_versionRegistry_WatchAllCurrent(t *testing.T) {
 	deleteCancelCtx, deleteCancelFunc := context.WithCancel(context.Background())
 	errorCancelCtx, errorCancelFunc := context.WithCancel(context.Background())
 	ldapRegistryCm := &corev1.ConfigMap{Data: map[string]string{"current": ldapVersionStr}, ObjectMeta: metav1.ObjectMeta{Labels: ldapVersionRegistryLabelMap, ResourceVersion: "1"}}
-	initialDoguVersionCtx := map[SimpleDoguName]core.Version{"ldap": parseVersionStr(t, ldapVersionStr)}
+	initialDoguVersionCtx := map[dogu.SimpleName]core.Version{"ldap": parseVersionStr(t, ldapVersionStr)}
 	casRegistryCm := &corev1.ConfigMap{Data: map[string]string{"current": casVersionStr}, ObjectMeta: metav1.ObjectMeta{Labels: casVersionRegistryLabelMap, ResourceVersion: "1"}}
 	registryCmList := &corev1.ConfigMapList{Items: []corev1.ConfigMap{*ldapRegistryCm}, ListMeta: metav1.ListMeta{ResourceVersion: "1"}}
 	emptyLdapRegistryCm := &corev1.ConfigMap{Data: map[string]string{}, ObjectMeta: metav1.ObjectMeta{Labels: ldapVersionRegistryLabelMap, ResourceVersion: "1"}}
@@ -435,7 +445,7 @@ func Test_versionRegistry_WatchAllCurrent(t *testing.T) {
 		configMapClientFn func(t *testing.T, watchInterface *watch.FakeWatcher) configMapClient
 		args              args
 		eventMockFn       func(watchInterface *watch.FakeWatcher)
-		expectFn          func(t *testing.T, watchCh <-chan CurrentVersionsWatchResult)
+		expectFn          func(t *testing.T, watchCh <-chan dogu.CurrentVersionsWatchResult)
 		wantErr           assert.ErrorAssertionFunc
 	}{
 		{
@@ -451,7 +461,7 @@ func Test_versionRegistry_WatchAllCurrent(t *testing.T) {
 				return assert.True(t, cloudoguerrors.IsGenericError(err), "error is not generic error", i) &&
 					assert.ErrorContains(t, err, "failed to list initial descriptor configmaps: failed to get all cluster native local dogu registries", i)
 			},
-			expectFn: func(t *testing.T, watchCh <-chan CurrentVersionsWatchResult) {
+			expectFn: func(t *testing.T, watchCh <-chan dogu.CurrentVersionsWatchResult) {
 				assert.Nil(t, watchCh)
 			},
 		},
@@ -470,7 +480,7 @@ func Test_versionRegistry_WatchAllCurrent(t *testing.T) {
 				return assert.True(t, cloudoguerrors.IsGenericError(err), "error is not generic error", i) &&
 					assert.ErrorContains(t, err, "failed to create persistence context for current dogu versions: failed to parse version \"abc\" for dogu \"cas\": failed to parse major version abc", i)
 			},
-			expectFn: func(t *testing.T, watchCh <-chan CurrentVersionsWatchResult) {
+			expectFn: func(t *testing.T, watchCh <-chan dogu.CurrentVersionsWatchResult) {
 				assert.Nil(t, watchCh)
 			},
 		},
@@ -489,7 +499,7 @@ func Test_versionRegistry_WatchAllCurrent(t *testing.T) {
 				return assert.True(t, cloudoguerrors.IsGenericError(err)) &&
 					assert.ErrorContains(t, err, "failed to create watch for current dogu versions")
 			},
-			expectFn: func(t *testing.T, watchCh <-chan CurrentVersionsWatchResult) {
+			expectFn: func(t *testing.T, watchCh <-chan dogu.CurrentVersionsWatchResult) {
 				assert.Nil(t, watchCh)
 			},
 		},
@@ -506,13 +516,14 @@ func Test_versionRegistry_WatchAllCurrent(t *testing.T) {
 			eventMockFn: func(watchInterface *watch.FakeWatcher) {
 				watchInterface.Add(casRegistryCm)
 			},
-			expectFn: func(t *testing.T, watchCh <-chan CurrentVersionsWatchResult) {
+			expectFn: func(t *testing.T, watchCh <-chan dogu.CurrentVersionsWatchResult) {
 				result := <-watchCh
 				require.NoError(t, result.Err)
 				assert.Equal(t, initialDoguVersionCtx, result.PrevVersions)
 				casVersion := parseVersionStr(t, casVersionStr)
-				assert.Equal(t, result.Versions, map[SimpleDoguName]core.Version{"ldap": parseVersionStr(t, ldapVersionStr), "cas": casVersion})
-				assert.Equal(t, []DoguVersion{{Name: "cas", Version: casVersion}}, result.Diff)
+				assert.Equal(t, result.Versions, map[dogu.SimpleName]core.Version{"ldap": parseVersionStr(t, ldapVersionStr), "cas": casVersion})
+				assert.Equal(t, []dogu.SimpleNameVersion{{
+					Name: "cas", Version: casVersion}}, result.Diff)
 
 				addCancelFunc()
 			},
@@ -533,13 +544,14 @@ func Test_versionRegistry_WatchAllCurrent(t *testing.T) {
 				// We have to send two events because is not possible to check if no event is thrown.
 				watchInterface.Add(casRegistryCm)
 			},
-			expectFn: func(t *testing.T, watchCh <-chan CurrentVersionsWatchResult) {
+			expectFn: func(t *testing.T, watchCh <-chan dogu.CurrentVersionsWatchResult) {
 				result := <-watchCh
 				require.NoError(t, result.Err)
 				assert.Equal(t, initialDoguVersionCtx, result.PrevVersions)
 				casVersion := parseVersionStr(t, casVersionStr)
-				assert.Equal(t, result.Versions, map[SimpleDoguName]core.Version{"ldap": parseVersionStr(t, ldapVersionStr), "cas": casVersion})
-				assert.Equal(t, []DoguVersion{{Name: "cas", Version: casVersion}}, result.Diff)
+				assert.Equal(t, result.Versions, map[dogu.SimpleName]core.Version{"ldap": parseVersionStr(t, ldapVersionStr), "cas": casVersion})
+				assert.Equal(t, []dogu.SimpleNameVersion{{
+					Name: "cas", Version: casVersion}}, result.Diff)
 
 				emptyAddCancelFunc()
 			},
@@ -559,13 +571,14 @@ func Test_versionRegistry_WatchAllCurrent(t *testing.T) {
 				configMap := &corev1.ConfigMap{Data: map[string]string{"current": upgradeLdapVersionStr}, ObjectMeta: metav1.ObjectMeta{Labels: ldapVersionRegistryLabelMap, ResourceVersion: "2"}}
 				watchInterface.Modify(configMap)
 			},
-			expectFn: func(t *testing.T, watchCh <-chan CurrentVersionsWatchResult) {
+			expectFn: func(t *testing.T, watchCh <-chan dogu.CurrentVersionsWatchResult) {
 				result := <-watchCh
 				require.NoError(t, result.Err)
 				assert.Equal(t, initialDoguVersionCtx, result.PrevVersions)
 				upgradedLdapVersion := parseVersionStr(t, upgradeLdapVersionStr)
-				assert.Equal(t, result.Versions, map[SimpleDoguName]core.Version{"ldap": upgradedLdapVersion})
-				assert.Equal(t, []DoguVersion{{Name: "ldap", Version: upgradedLdapVersion}}, result.Diff)
+				assert.Equal(t, result.Versions, map[dogu.SimpleName]core.Version{"ldap": upgradedLdapVersion})
+				assert.Equal(t, []dogu.SimpleNameVersion{{
+					Name: "ldap", Version: upgradedLdapVersion}}, result.Diff)
 
 				modifyCancelFunc()
 			},
@@ -585,13 +598,14 @@ func Test_versionRegistry_WatchAllCurrent(t *testing.T) {
 				object := &corev1.ConfigMap{Data: map[string]string{"current": ldapVersionStr}, ObjectMeta: metav1.ObjectMeta{Labels: ldapVersionRegistryLabelMap, ResourceVersion: "2"}}
 				watchInterface.Delete(object)
 			},
-			expectFn: func(t *testing.T, watchCh <-chan CurrentVersionsWatchResult) {
+			expectFn: func(t *testing.T, watchCh <-chan dogu.CurrentVersionsWatchResult) {
 				result := <-watchCh
 				require.NoError(t, result.Err)
 				assert.Equal(t, initialDoguVersionCtx, result.PrevVersions)
 				ldapVersion := parseVersionStr(t, ldapVersionStr)
-				assert.Equal(t, result.Versions, map[SimpleDoguName]core.Version{})
-				assert.Equal(t, []DoguVersion{{Name: "ldap", Version: ldapVersion}}, result.Diff)
+				assert.Equal(t, result.Versions, map[dogu.SimpleName]core.Version{})
+				assert.Equal(t, []dogu.SimpleNameVersion{{
+					Name: "ldap", Version: ldapVersion}}, result.Diff)
 
 				deleteCancelFunc()
 			},
@@ -611,7 +625,7 @@ func Test_versionRegistry_WatchAllCurrent(t *testing.T) {
 				watchInterface.Error(&metav1.Status{Status: "123", Message: "message"})
 				errorCancelFunc()
 			},
-			expectFn: func(t *testing.T, watchCh <-chan CurrentVersionsWatchResult) {
+			expectFn: func(t *testing.T, watchCh <-chan dogu.CurrentVersionsWatchResult) {
 				result := <-watchCh
 				require.NoError(t, result.Err)
 			},
@@ -756,7 +770,7 @@ func Test_handleEvent(t *testing.T) {
 			Object: &metav1.Status{},
 		}
 
-		channel := make(chan CurrentVersionsWatchResult)
+		channel := make(chan dogu.CurrentVersionsWatchResult)
 
 		// when
 		go handleEvent(testCtx, event, nil, channel)
@@ -776,7 +790,7 @@ func Test_handleEvent(t *testing.T) {
 			Object: nil,
 		}
 
-		channel := make(chan CurrentVersionsWatchResult)
+		channel := make(chan dogu.CurrentVersionsWatchResult)
 
 		// when
 		go handleEvent(testCtx, event, nil, channel)
@@ -818,7 +832,7 @@ func Test_handleModifiedWatchEvent(t *testing.T) {
 			Object: noCurrentConfigMap,
 		}
 
-		persistentContext := make(map[SimpleDoguName]core.Version)
+		persistentContext := make(map[dogu.SimpleName]core.Version)
 
 		// when
 		err := handleModifiedWatchEvent(testCtx, event, persistentContext, nil)
@@ -836,9 +850,9 @@ func Test_handleModifiedWatchEvent(t *testing.T) {
 			Object: noCurrentConfigMap,
 		}
 
-		persistentContext := map[SimpleDoguName]core.Version{"ldap": parseVersionStr(t, "1.0.0")}
-		expectedOldVersions := map[SimpleDoguName]core.Version{"ldap": parseVersionStr(t, "1.0.0")}
-		channel := make(chan CurrentVersionsWatchResult)
+		persistentContext := map[dogu.SimpleName]core.Version{"ldap": parseVersionStr(t, "1.0.0")}
+		expectedOldVersions := map[dogu.SimpleName]core.Version{"ldap": parseVersionStr(t, "1.0.0")}
+		channel := make(chan dogu.CurrentVersionsWatchResult)
 
 		// when
 		go func() {
@@ -851,7 +865,8 @@ func Test_handleModifiedWatchEvent(t *testing.T) {
 		require.NoError(t, result.Err)
 		assert.Len(t, result.Versions, 0)
 		assert.Equal(t, result.PrevVersions, expectedOldVersions)
-		assert.Equal(t, result.Diff, []DoguVersion{{"ldap", parseVersionStr(t, "1.0.0")}})
+		assert.Equal(t, result.Diff, []dogu.SimpleNameVersion{{
+			Name: "ldap", Version: parseVersionStr(t, "1.0.0")}})
 		assert.Len(t, persistentContext, 0)
 	})
 }